Type encoders are no longer stored as shared pointers. Added support for 1-byte skip offsets.

+++ tMemoryBuffer.h
@@ -253,7 +253,7 @@ public:
   /*!
    * \return Buffer size
    */
-  inline int GetSize() const
+  inline size_t GetSize() const
   {
     return cur_size;
   }
